refactor: proguard keep 범위 좁힘 + Optional storage/model 로 이동
- Optional 도 Moshi 직렬화 대상이므로 storage/model 로 이동하여 storage.model.** 한 패키지에 모든 직렬화 모델을 모음
- proguard 변경:
  - 제거: com.wafflestudio.snutt2.** 전체 enum keep (storage.model 안의 enum 만 keep 필요)
  - 제거: lib.android.NetworkLog (이미 storage.model 로 이동됨)
  - 제거: domain.model.** (도메인 모델이 더 이상 storage 직렬화 대상이 아님)
  - 좁힘: storage.** → storage.model.**
  - 유지: network.** (Retrofit DTO 는 그대로 reflection 필요)
- Optional import 가 들어있던 6개 파일 (SNUTTStorage, PrefCacheImpl, PrefValueMetaData, 4개 Repository) import 경로 갱신
- minifyStagingReleaseWithR8 통과로 R8 가 좁힌 rule 로 빌드 완료됨을 검증
