+++ app/src/main/java/com/wafflestudio/snutt2/data/semesterstatus/SemesterStatusRepositoryImpl.kt
@@ -8,7 +8,7 @@ import com.wafflestudio.snutt2.network.error.toDomainError
 import com.wafflestudio.snutt2.storage.SNUTTStorage
 import com.wafflestudio.snutt2.storage.model.toDomainModel
 import com.wafflestudio.snutt2.storage.model.toLocalEntity
-import com.wafflestudio.snutt2.storage.toOptional
+import com.wafflestudio.snutt2.storage.model.toOptional
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
