+++ app/src/main/java/com/wafflestudio/snutt2/storage/SNUTTStorage.kt
@@ -1,6 +1,7 @@
 package com.wafflestudio.snutt2.storage
 
 import com.wafflestudio.snutt2.storage.model.NetworkLog
+import com.wafflestudio.snutt2.storage.model.Optional
 import com.wafflestudio.snutt2.storage.model.SemesterStatusLocalEntity
 import com.wafflestudio.snutt2.storage.model.SimpleTableLocalEntity
 import com.wafflestudio.snutt2.storage.model.TableLectureCustomData
